% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame.R
\name{pv_flame_condense}
\alias{pv_flame_condense}
\title{Condensed flame graph}
\usage{
pv_flame_condense(x, n = 10, width = 50)
}
\arguments{
\item{x}{A profvis object.}

\item{n}{Number of hot paths to show.}

\item{width}{Width of bars.}
}
\value{
Invisibly returns a \code{debrief_flame_condense} object. Use
\code{capture.output()} to capture the formatted text output.
}
\description{
Shows a simplified, condensed view of the flame graph focusing on the
hottest paths.
}
\examples{
p <- pv_example()
pv_flame_condense(p)
}
