% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antipatterns.R
\name{pv_print_antipatterns}
\alias{pv_print_antipatterns}
\title{Print anti-patterns analysis}
\usage{
pv_print_antipatterns(x)
}
\arguments{
\item{x}{A profvis object.}
}
\value{
Invisibly returns the anti-patterns data frame.
}
\description{
Print anti-patterns analysis
}
