% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-depth.R
\name{pv_print_call_depth}
\alias{pv_print_call_depth}
\title{Print call depth breakdown}
\usage{
pv_print_call_depth(x)
}
\arguments{
\item{x}{A profvis object.}
}
\value{
Invisibly returns a \code{debrief_call_depth} object. Use
\code{capture.output()} to capture the formatted text output.
}
\description{
Print call depth breakdown
}
\examples{
p <- pv_example()
pv_print_call_depth(p)

}
