% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-stats.R
\name{pv_print_call_stats}
\alias{pv_print_call_stats}
\title{Print call statistics}
\usage{
pv_print_call_stats(x, n = 20)
}
\arguments{
\item{x}{A profvis object.}

\item{n}{Number of functions to show.}
}
\value{
Invisibly returns a \code{debrief_call_stats} object. Use
\code{capture.output()} to capture the formatted text output.
}
\description{
Print call statistics
}
\examples{
p <- pv_example()
pv_print_call_stats(p)

}
