% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestions.R
\name{pv_print_suggestions}
\alias{pv_print_suggestions}
\title{Print optimization suggestions}
\usage{
pv_print_suggestions(x)
}
\arguments{
\item{x}{A profvis object.}
}
\value{
Invisibly returns a \code{debrief_suggestions} object. Use
\code{capture.output()} to capture the formatted text output.
}
\description{
Print optimization suggestions
}
\examples{
p <- pv_example("gc")
pv_print_suggestions(p)

}
