% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-stats.R
\name{pv_print_call_stats}
\alias{pv_print_call_stats}
\title{Print call statistics}
\usage{
pv_print_call_stats(x, n = 20)
}
\arguments{
\item{x}{A profvis object.}

\item{n}{Number of functions to show.}
}
\value{
Invisibly returns the call stats data frame.
}
\description{
Print call statistics
}
