% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-summary.R
\name{pv_print_file_summary}
\alias{pv_print_file_summary}
\title{Print file summary}
\usage{
pv_print_file_summary(x)
}
\arguments{
\item{x}{A profvis object.}
}
\value{
Invisibly returns the file summary data frame.
}
\description{
Print file summary
}
