% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antipatterns.R
\name{pv_print_gc_pressure}
\alias{pv_print_gc_pressure}
\title{Print GC pressure analysis}
\usage{
pv_print_gc_pressure(x)
}
\arguments{
\item{x}{A profvis object.}
}
\value{
Invisibly returns the GC pressure data frame.
}
\description{
Print GC pressure analysis
}
\examples{
p <- pv_example("gc")
pv_print_gc_pressure(p)

}
