% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot-paths.R
\name{pv_print_hot_paths}
\alias{pv_print_hot_paths}
\title{Print hot paths in readable format}
\usage{
pv_print_hot_paths(x, n = 10, include_source = TRUE)
}
\arguments{
\item{x}{A profvis object.}

\item{n}{Number of paths to show.}

\item{include_source}{Include source references in output.}
}
\value{
Invisibly returns the hot paths data frame.
}
\description{
Print hot paths in readable format
}
