% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive.R
\name{pv_print_recursive}
\alias{pv_print_recursive}
\title{Print recursive functions analysis}
\usage{
pv_print_recursive(x)
}
\arguments{
\item{x}{A profvis object.}
}
\value{
Invisibly returns the recursive functions data frame.
}
\description{
Print recursive functions analysis
}
